simpleTwit_backend/requirement.txt
flask
flask_cors
flask_sqlalchemy
sqlalchemy
flask_login
flask_bcrypt
sqlalchemy_serializer